Hacking in in-process timeouts for JUnit tests. I'm not particularly happy about having to add this, since it means we have two different places where timeouts can happen - from the host process in Driver.java and in the target process.
One possible way to remove the redundancy is to make the clientside TestRunner responsible for committing suicide if the process hangs for too long. The host process would need to hand across the timeouts, but that's not too bad, and probably much better than the hard coded  one minute timouts for JUnit test methods that I've added here.