Fixed vogar to work without /sdcard and with new bouncycastle classpath
Replaced --device-runner-dir with --device-dir, which still defaults
to /sdcard/vogar.  Run ouput is now in a subdir
<device-dir>/vogar/run. This is so it can be removed separately from
<device-dir>/md5-cache, which is no longer hardwired to
/sdcard/tmp/vogar-md5-cache/

Also updated the build and runtime classpath to include
bouncycastle. Rather than do this in two places, fixed to have one
definition in AndroidSdk of the BOOTCLASSPATH jars. The build and
runtime classpaths were divergent from each other and from the real
device $BOOTCLASSPATH so I brought this into alignment with the
device, or rather system/core/rootdir/init.rc.