Disable JCA security checks on the RI
This allows the running of tests that insert their own Provider class. This
was failing silently before, but replacing this one field with a Map that
returns true bypasses the checks.

This should probably log an error when it can't find the class via reflection
on platforms where it expects to find it.